glue statistics endpoint together
